requirements.txt
numpy==1.24.1
opencv-python==4.7.0.68
paho-mqtt==1.6.1
python-dotenv==0.21.0
